backend/requirements/base.txt
Django==6.0.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-filter==25.2
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
celery==5.6.2
redis==7.1.0
django-celery-beat==2.8.1
python-dateutil==2.9.0.post0
pytz==2025.2
Pillow==12.1.0
pillow-heif==1.1.1
qrcode==8.2
requests==2.32.5
matplotlib==3.10.8
django-storages==1.14.6
boto3==1.42.60
exponent-server-sdk==2.2.0
python-dotenv==1.2.1
python-magic-bin==0.4.14
openpyxl==3.1.5
weasyprint==68.1
